+++ src/RcppExports.cpp
@@ -12,25 +12,25 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // Rcpp_delaunay
-Rcpp::List Rcpp_delaunay(const arma::mat& points);
-RcppExport SEXP _RCDT_Rcpp_delaunay(SEXP pointsSEXP) {
+arma::umat Rcpp_delaunay(const arma::mat& tpoints);
+RcppExport SEXP _RCDT_Rcpp_delaunay(SEXP tpointsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const arma::mat& >::type points(pointsSEXP);
-    rcpp_result_gen = Rcpp::wrap(Rcpp_delaunay(points));
+    Rcpp::traits::input_parameter< const arma::mat& >::type tpoints(tpointsSEXP);
+    rcpp_result_gen = Rcpp::wrap(Rcpp_delaunay(tpoints));
     return rcpp_result_gen;
 END_RCPP
 }
 // Rcpp_constrained_delaunay
-Rcpp::List Rcpp_constrained_delaunay(const arma::mat& points, const arma::umat& edges);
-RcppExport SEXP _RCDT_Rcpp_constrained_delaunay(SEXP pointsSEXP, SEXP edgesSEXP) {
+Rcpp::List Rcpp_constrained_delaunay(const arma::mat& tpoints, const arma::umat& tedges);
+RcppExport SEXP _RCDT_Rcpp_constrained_delaunay(SEXP tpointsSEXP, SEXP tedgesSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const arma::mat& >::type points(pointsSEXP);
-    Rcpp::traits::input_parameter< const arma::umat& >::type edges(edgesSEXP);
-    rcpp_result_gen = Rcpp::wrap(Rcpp_constrained_delaunay(points, edges));
+    Rcpp::traits::input_parameter< const arma::mat& >::type tpoints(tpointsSEXP);
+    Rcpp::traits::input_parameter< const arma::umat& >::type tedges(tedgesSEXP);
+    rcpp_result_gen = Rcpp::wrap(Rcpp_constrained_delaunay(tpoints, tedges));
     return rcpp_result_gen;
 END_RCPP
 }