+++ NEWS.md
@@ -1,3 +1,12 @@
+# RCDT 1.2.0 (2022-08-07)
+
+* Speed improvement, by removing some useless code. 
+
+* Upgraded CDT.
+
+* Handle intersecting constraint edges.
+
+
 # RCDT 1.1.0 (2022-04-06)
 
 * New example to show that the triangulation depends on the order of the points.